:recycle: RegisterGifticonInfoMessageWrapperView 제거  (#35)
- RegisterGifticonView 내부에 fileprivate으로 InfoMessageView 추가하여 대체
